samples: matter: Fixed an issue with pwm overflow
A uint32 variable had been overflowed when
The new PWM value for LED was set.

+++ samples/matter/common/src/pwm_device.cpp
@@ -99,5 +99,5 @@ void PWMDevice::UpdateLight()
 	const uint8_t effectiveLevel =
 		mState == kState_On ? chip::min<uint8_t>(mLevel - mMinLevel, maxEffectiveLevel) : 0;
 
-	pwm_set_pulse_dt(mPwmDevice, mPwmDevice->period * effectiveLevel / maxEffectiveLevel);
+	pwm_set_pulse_dt(mPwmDevice, static_cast<uint32_t>(static_cast<const uint64_t>(mPwmDevice->period) * effectiveLevel / maxEffectiveLevel));
 }
